Fix ambiguous matching when removing previous links The regex pattern that is being used here has an issue where, because of the fact that square braces [] are also used to display the difficulty of a level, it will match with the starting brace of the difficulty string instead of the brace that's supposed to be the beginning of the hyperlinked text. That's because this regex rule doesn't care what or how many characters are between the starting and closing brace; as long as there's a closing brace, it's a match.
The new regex rule is less ambiguous, as there must be a strictly numeric value between the braces for it to match (difficulty string has an "x", which means it won't match). I don't know how the bot formats strings that are whole integers, so the decimal point and places are optional.